Refactored breakpoint ops to their own file/class
